fetaure/SCAT 980 - Publishing RFI to suppliers
