+++ src/main/features/rfi/controller/responsedate.ts
@@ -16,6 +16,8 @@ export const GET_RESPONSE_DATE = async (req: express.Request, res: express.Respo
 
 export const POST_RESPONSE_DATE = async (req: express.Request, res: express.Response) => {
    const RequestBodyValues = Object.values(req.body);
+   const {supplier_period_for_clarification_period} = req.body;
+   req.session['endDate'] = supplier_period_for_clarification_period;
    const filterWithQuestions = RequestBodyValues.map(aQuestions => {
       const anEntry = aQuestions.split("*");
       return { Question: anEntry[0], value: anEntry[1] };
@@ -45,16 +47,16 @@ export const POST_RESPONSE_DATE = async (req: express.Request, res: express.Resp
          criterianStorage.push(rebased_object_with_requirements)
       }
       criterianStorage = criterianStorage.flat();
+      const Criterian_ID = criterianStorage[0].criterianId;
       criterianStorage = criterianStorage.filter(AField => AField.OCDS.id === keyDateselector)
-      const apiData_baseURL = `/tenders/projects/${proc_id}/events/${event_id}/criteria/Criterion 2/groups/${keyDateselector}/questions`;
+      const apiData_baseURL = `/tenders/projects/${proc_id}/events/${event_id}/criteria/${Criterian_ID}/groups/${keyDateselector}/questions`;
       const fetchQuestions = await TenderApi.Instance(SESSION_ID).get(apiData_baseURL);
       const fetchQuestionsData = fetchQuestions.data;
       const allunfilledAnswer = fetchQuestionsData.filter(anAswer => anAswer.nonOCDS.options.length == 0).map(aQuestion => aQuestion.OCDS.id)
-      console.log(allunfilledAnswer)
       for (const answers of allunfilledAnswer) {
          const proc_id = req.session.projectId;
          const event_id = req.session.eventId;
-         const id = "Criterion 2";
+         const id = Criterian_ID;
          const group_id = "Key Dates";
          const question_id = answers;
          const findFilterQuestion = filterWithQuestions.filter(question => question.Question === question_id);
@@ -73,10 +75,11 @@ export const POST_RESPONSE_DATE = async (req: express.Request, res: express.Resp
          const answerBaseURL = `/tenders/projects/${proc_id}/events/${event_id}/criteria/${id}/groups/${group_id}/questions/${question_id}`;
          await TenderApi.Instance(SESSION_ID).put(answerBaseURL, answerBody);
       }
-      const response = await TenderApi.Instance(SESSION_ID).put(`journeys/${event_id}/steps/13`, 'Completed');
+       const response = await TenderApi.Instance(SESSION_ID).put(`journeys/${event_id}/steps/13`, 'Completed');
       if (response.status == HttpStatusCode.OK) {
          await TenderApi.Instance(SESSION_ID).put(`journeys/${event_id}/steps/14`, 'Not started');
       }
+      
       res.redirect('/rfi/review')
    } catch (error) {
       LoggTracer.errorLogger(res, error, `${req.headers.host}${req.originalUrl}`, null,
@@ -98,8 +101,6 @@ export const POST_ADD_RESPONSE_DATE = async (req: express.Request, res: express.
       clarification_date_hourFormat,
       selected_question_id
    } = req.body;
-
-   console.log(req.body)
 
    clarification_date_day = Number(clarification_date_day);
    clarification_date_month = Number(clarification_date_month);
@@ -148,11 +149,30 @@ export const POST_ADD_RESPONSE_DATE = async (req: express.Request, res: express.
       try {
          const proc_id = req.session.projectId;
          const event_id = req.session.eventId;
-         const id = "Criterion 2";
+       
          const group_id = "Key Dates";
          const question_id = selected_question_id;
 
-
+         let baseURL = `/tenders/projects/${proc_id}/events/${event_id}`;
+         baseURL = baseURL + '/criteria'
+         const fetch_dynamic_api = await TenderApi.Instance(SESSION_ID).get(baseURL);
+         const fetch_dynamic_api_data = fetch_dynamic_api?.data;
+         const extracted_criterion_based = fetch_dynamic_api_data?.map((criterian) => criterian?.id);
+         let criterianStorage = [];
+         for (const aURI of extracted_criterion_based) {
+            const criterian_bas_url = `/tenders/projects/${proc_id}/events/${event_id}/criteria/${aURI}/groups`;
+            const fetch_criterian_group_data = await TenderApi.Instance(SESSION_ID).get(criterian_bas_url);
+            const criterian_array = fetch_criterian_group_data?.data;
+            const rebased_object_with_requirements = criterian_array?.map((anItem) => {
+               const object = anItem;
+               object['criterianId'] = aURI;
+               return object;
+            })
+            criterianStorage.push(rebased_object_with_requirements)
+         }
+         criterianStorage = criterianStorage.flat();
+         const Criterian_ID = criterianStorage[0].criterianId;
+         const id = Criterian_ID;
          const answerBaseURL = `/tenders/projects/${proc_id}/events/${event_id}/criteria/${id}/groups/${group_id}/questions/${question_id}`;
          await TenderApi.Instance(SESSION_ID).put(answerBaseURL, answerBody);
          res.redirect('/rfi/response-date')
