+++ src/main/features/rfi/helpers/responsedate.ts
@@ -49,9 +49,10 @@ export const RESPONSEDATEHELPER = async (req: express.Request, res: express.Resp
         }
 
         criterianStorage = criterianStorage.flat();
-        criterianStorage = criterianStorage.filter(AField => AField.OCDS.id === keyDateselector)
+        criterianStorage = criterianStorage.filter(AField => AField.OCDS.id === keyDateselector);
+        const Criterian_ID = criterianStorage[0].criterianId;
         const prompt = criterianStorage[0].nonOCDS.prompt;
-        const apiData_baseURL = `/tenders/projects/${proc_id}/events/${event_id}/criteria/Criterion 2/groups/${keyDateselector}/questions`;
+        const apiData_baseURL = `/tenders/projects/${proc_id}/events/${event_id}/criteria/${Criterian_ID}/groups/${keyDateselector}/questions`;
         const fetchQuestions = await DynamicFrameworkInstance.Instance(SESSION_ID).get(apiData_baseURL);
         let fetchQuestionsData = fetchQuestions.data;
         const rfi_clarification_date = new Date().toLocaleDateString('en-uk', { weekday: "long", year: "numeric", month: "short", day: "numeric" })
